Translates documents appropriately. Fixes lines. Adds dark text as color for Text. Tries to fix translation of presets (to no avail).
