+++ src/renderer/src/Theme.ts
@@ -3,11 +3,14 @@ import { createTheme } from '@mui/material/styles'
 import {
 	COMAPEO_BLUE,
 	DARK_COMAPEO_BLUE,
+	DARK_GREY,
 	DARK_ORANGE,
+	DARK_TEXT,
 	GREEN,
 	LIGHT_COMAPEO_BLUE,
 	ORANGE,
 	RED,
+	WHITE,
 } from './colors'
 
 const theme = createTheme({
@@ -19,12 +22,15 @@ const theme = createTheme({
 		},
 		body2: {
 			fontSize: '0.875rem',
+			color: DARK_TEXT,
 		},
 		subtitle1: {
 			fontSize: '1.125rem',
+			color: DARK_TEXT,
 		},
 		subtitle2: {
 			fontSize: '1rem',
+			color: DARK_TEXT,
 		},
 		button: {
 			fontSize: '1rem',
@@ -33,17 +39,24 @@ const theme = createTheme({
 		},
 		caption: {
 			fontSize: '0.75rem',
+			color: DARK_TEXT,
 		},
 		h1: {
 			fontSize: '2rem',
 			fontWeight: 700,
+			color: DARK_TEXT,
 		},
 		h2: {
 			fontSize: '1.5rem',
 			fontWeight: 600,
+			color: DARK_TEXT,
 		},
 	},
 	palette: {
+		text: {
+			primary: DARK_TEXT,
+			secondary: DARK_GREY,
+		},
 		primary: {
 			main: COMAPEO_BLUE,
 			dark: DARK_COMAPEO_BLUE,
@@ -58,6 +71,9 @@ const theme = createTheme({
 		},
 		error: {
 			main: RED,
+		},
+		background: {
+			default: WHITE,
 		},
 	},
 	components: {
