+++ src/renderer/src/components/Observations/ObservationListItem.tsx
@@ -1,16 +1,12 @@
 import type { Observation } from '@comapeo/schema'
 import { styled } from '@mui/material/styles'
-import {
-	FormattedDate,
-	FormattedTime,
-	defineMessages,
-	useIntl,
-} from 'react-intl'
+import { FormattedDate, FormattedTime } from 'react-intl'
 
 import { VERY_LIGHT_GREY } from '../../colors'
 import { useObservationWithPreset } from '../../hooks/useObservationWithPreset'
 import { PresetCircleIcon } from '../PresetCircleIcon'
 import { Text } from '../Text'
+import { FormattedPresetName } from '../formattedData'
 
 type Props = {
 	observation: Observation
@@ -44,21 +40,11 @@ const PhotoContainer = styled('img')({
 	objectFit: 'cover',
 })
 
-const m = defineMessages({
-	observation: {
-		// Keep id stable for translations
-		id: 'screens.Observation.ObservationView.observation',
-		defaultMessage: 'Observation',
-		description: 'Default name of observation with no matching preset',
-	},
-})
-
 export function ObservationListItem({
 	observation,
 	projectId,
 	onClick,
 }: Props) {
-	const { formatMessage } = useIntl()
 	const preset = useObservationWithPreset(observation, projectId ?? '')
 	const createdAt = observation.createdAt
 		? new Date(observation.createdAt)
@@ -67,17 +53,13 @@ export function ObservationListItem({
 	const photoAttachment = observation.attachments.find(
 		(att) => att.type === 'photo',
 	)
-	const displayName = preset
-		? formatMessage({
-				id: `presets.${preset.docId}.name`,
-				defaultMessage: preset.name,
-			})
-		: formatMessage(m.observation)
 
 	return (
 		<Container onClick={onClick}>
 			<TextContainer>
-				<Text style={{ fontWeight: 500 }}>{displayName}</Text>
+				<Text style={{ fontWeight: 500 }}>
+					<FormattedPresetName preset={preset} />
+				</Text>
 				<Text style={{ fontSize: 10, fontWeight: 400 }}>
 					<FormattedDate
 						value={createdAt}
