+++ src/renderer/src/hooks/useObservationWithPreset.ts
@@ -1,6 +1,7 @@
 import { useMemo } from 'react'
 import { useManyDocs } from '@comapeo/core-react'
 import type { Observation } from '@comapeo/schema'
+import { useIntl } from 'react-intl'
 
 import { matchPreset } from '../lib/matchPreset'
 
@@ -8,11 +9,12 @@ export function useObservationWithPreset(
 	observation: Observation,
 	projectId: string,
 ) {
+	const { locale } = useIntl()
 	const { data: presets = [] } = useManyDocs({
 		projectId: projectId || '',
 		docType: 'preset',
 		includeDeleted: false,
-		lang: 'en',
+		lang: locale,
 	})
 
 	const matchedPreset = useMemo(() => {
